refactor(media): streamline media services and consolidate firebase logic
MediaUploadService now directly injects and uses FirebaseStorageStrategy,
removing the need for a generic strategy parameter in its upload method.
MediaDbService similarly injects FirebaseDbStrategy directly.
Firebase-specific helper methods (buildStoragePath, generateSignedUrl)
have been moved from MediaUploadService to FirebaseStorageStrategy
for better encapsulation and cohesion of Firebase storage logic.
Adds UPLOAD_SOUNDSHOT to PipelineType enum, enabling audio uploads.
