+++ apps/api/src/common/pipeline/types.ts
@@ -2,4 +2,5 @@
 export enum PipelineType {
   UPLOAD_IMAGE = 'upload → image',
   UPLOAD_VIDEO = 'upload → video',
+  UPLOAD_SOUNDSHOT = 'upload → audio',
 }
